finish collect song

+++ main.py
@@ -1,5 +1,6 @@
 import os
 import random
+import time
 import discord
 from discord.ext import commands
 import pandas as pd
@@ -10,48 +11,25 @@ client = commands.Bot(command_prefix='$')
 
 @client.event
 async def on_ready():
-    print("I'm online now!!")
-    print(" --------------")
+    print("I'm ready now!!")
+    print("---------------")
 
 
-# @client.command()
-# async def hello(ctx):
-#     await ctx.send('Hello!')
-
-
-@client.command(pass_context=True)
-async def join(ctx):
-    if (ctx.author.voice):
-        channel = ctx.message.author.voice.channel
-        await channel.connect()
-    else:
-        await ctx.send('You are not in the channel!')
-
-
-@client.command(pass_context=True)
-async def leave(ctx):
-    if (ctx.voice_client):
-        await ctx.guild.voice_client.disconnect()
-    else:
-        await ctx.send("I'm not in the channel!!")
-
-
-# collect messages
 @client.event
 async def on_message(message):
     if message.author == client.user:
         return
-    elif message.content.startswith('_'):
+    elif message.content.startswith('$'):
 
-        cmd = message.content.split()[0].replace("_", "")
+        cmd = message.content.split()[0].replace("$", "")
         if len(message.content.split()) > 1:
             parameters = message.content.split()[1:]
 
         # Bot Commands
 
-        if cmd == 'scan':
+        if cmd == 'collect':  # collect message
 
-            data = pd.DataFrame(columns=['content', 'time', 'author'])
+            data = pd.DataFrame(columns=['content'])
 
             # Acquiring the channel via the bot command
             if len(message.channel_mentions) > 0:
@@ -59,50 +37,43 @@ async def on_message(message):
             else:
                 channel = message.channel
 
-            # Aquiring the number of messages to be scraped via the bot command
+            # check number of song to collect
             if (len(message.content.split()) > 1):
-                for parameter in parameters:
-                    if parameter == "help":
-                        answer = discord.Embed(title="Command Format",
-                                               description="""`_scan <number_of_messages>`\n\n`<number_of_messages>` : **the number of messages you wish to scan**\n\n""",
-                                               colour=0x1a7794)
-                        await message.channel.send(embed=answer)
-                        return
-                    else:
-                        limit = int(parameter)
-            else:
+                limit = int(parameters[0])
+            else:  # set default at 100
                 limit = 100
 
-            answer = discord.Embed(title="Creating your Message History Dataframe",
+            answer = discord.Embed(title="Creating Sasa's song History Dataframe",
                                    description="Please Wait. The data will be sent to you privately once it's finished.",
                                    colour=0x1a7794)
 
             await message.channel.send(embed=answer)
 
-            def is_command(message):
-                if len(msg.content) == 0:
-                    return False
-                elif msg.content.split()[0] == '_scan':
+            def is_rythm_command(msg):
+                if msg.content[:2] == '!p':
                     return True
-                else:
-                    return False
+
+            def is_sasa_command(msg):
+                if msg.author.name == 'Nalinsa':
+                    return True
 
             # The added 1000 is so in case it skips messages for being
             async for msg in channel.history(limit=limit + 1000):
-                # a command or a message it sent, it will still read the
-                if msg.author != client.user:
-                    # the total amount originally specified by the user.
-                    if not is_command(msg):
-                        data = data.append({'content': msg.content,
-                                            'time': msg.created_at,
-                                            'author': msg.author.name}, ignore_index=True)
+                # collect songs that open from Sasa
+                if msg.author != client.user and is_rythm_command(msg) and is_sasa_command(msg):
+                    data = data.append(
+                        {'content': msg.content, }, ignore_index=True)
                     if len(data) == limit:
                         break
 
             # Turning the pandas dataframe into a .csv file and sending it to the user
-
             # Determining file name and location
-            file_location = f"{str(channel.guild.id) + '_' + str(channel.id)}.csv"
+            if message.author.name == '𝓧𝔂𝓷𝔃':
+                file_location = "./data/{}_data.csv".format(
+                    time.asctime().replace(' ', '_')).replace(':', '.')
+            else:
+                file_location = "data.csv"
+
             data.to_csv(file_location)  # Saving the file as a .csv via pandas
 
             answer = discord.Embed(title="Here is your .CSV File",
@@ -112,7 +83,14 @@ async def on_message(message):
             await message.author.send(embed=answer)
             # Sending the file
             await message.author.send(file=discord.File(file_location, filename='data.csv'))
-            os.remove(file_location)  # Deleting the file
+
+            if message.author.name != '𝓧𝔂𝓷𝔃':  # if message from me dont delete
+                os.remove(file_location)  # Deleting the file
+
+        elif cmd == 'pick':  # pick random song
+            return
+
     await client.process_commands(message)  # allow to use client.command
+
 
 client.run(TOKEN)
